feat: criacao de testes para a api

+++ api/Models/Funcionarios/Funcionario.cs
@@ -8,7 +8,7 @@ namespace api.Models.Funcionario
     public class Funcionario
     {
         public decimal Salario { get; set; }
-        public decimal PercentualAumento { get; set; }
+        public int PercentualAumento { get; set; }
         public decimal ValorAumento { get; set; }
         public decimal SalarioPosReajuste { get; set; }
     }
